Remove JobContext class dependency
Hadoop v2.0.4 changed JobContext from a class to an interface, which
means that a rubydoop.jar compiled against an old version cannot be
used in the new version if any methods on this class is executed.

See https://github.com/iconara/rubydoop/issues/9

This resolves the issue by duplicating the code in the mapper and
the reducer, as Mapper.Context and Reducer.Context are still classes.