Support default gems with no directory under gems
Rubygems 2.1 introduced a default gem feature and JRuby >= 1.7.5 uses
that feature to bundle some gems within the standard library gem path.

As the gems are bundled inside jruby-complete.jar, we do not need to add
them once more to the jar file. Trying to do so actually produces an
exception, since there is no gem directory for them.

This should resolve issue #15.