Add support for custom input formats
This adds an InputFormatProxy along the same lines as the other proxies.

Unfortunately, I couldn’t really figure out how to write a meaningful
unit test, so I implemented a (rather contrived) custom input format in
the integration tests to at least make sure the code runs.

A limitation is that the splits returned from the input format must be
standard Hadoop splits, as any input split must be serialisable and
that would require a significant amount of code.