Dynamically get JobContext configuration
As JobContext changed from class to interface between Hadoop 1 and 2,
we get an IncompatibleClassChangeError if we try to use the proxy class
compiled to either of the versions with the runtime of the other.

By getting the configuration method dynamically, we ensure that you can
compile the proxy class with either Hadoop version, and independent of
version used, it should still work together with either runtime.