Avoid ScriptingContainer#get("Constant")
For some unknown reason, it does not seem to work when the Java-code is
executed from within JRuby. In that context, the get call always returns
null (unless a thread-safe scripting container is created). The thread-
safe scripting container does not work at all from within Hadoop, as
variables sharing between threads becomes difficult.

Any instance of Kernel was replaced with null, as the default receiver
(main) has Kernel included, and thus should work exactly the same way.

For other constants, the get call was replaced with the more expensive
alternative runScriptlet which seems to work independent of context.