Add support for parallel job execution
This introduces `parallel` blocks into the configuration dsl to allow
jobs that should be executed in parallel. In addition, this adds
support for nested `sequence` blocks inside the `parallel` blocks in
order to be able to specify sequences of jobs to be executed in
separate swim-lanes.

Unfortunately, there is no way to listen for completion events for the
Hadoop Job class, which means the only simple way to implement support
for this nesting is by dispatching parallel jobs using threads. The
thread overhead should probably be acceptable, though, so lets not worry
about that until it becomes a real problem.