Don't require setup script in InstanceContainer
I accidentally forgot to set the JOB_SETUP_SCRIPT_KEY property in my
Ruby runner and found that it was required for the InstanceContainer to
load classes. However, completely removing it still seems to work. My
guess was that I wouldn't be able to find the classes only required in
the setup script, but it appears I was wrong.