Use puck-1.1.0 to include lib/rubydoop.jar
puck-1.1.0 supports specifying where in the JAR to put dependencies,
which is leveraged to place rubydoop.jar where Hadoop can find it on the
task nodes.

It might be uneccessary to put rubydoop.jar inside the lib direcory of
rubydoop, now that Hadoop should be able to find it anyhow. I haven't
really thought about the implications though.